requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
anthropic>=0.7.0
python-dotenv>=0.21.0
pydantic>=2.0.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiosmtplib>=2.0.0
email-validator>=2.0.0
pytest>=7.3.1
httpx>=0.24.1
python-jose[cryptography]>=3.3.0
